Finished POC charselect
