+++ Lang/EN.lang
@@ -302,3 +302,4 @@ http_509=Your request cannot be performed due to bandwidth limitation, report th
 http_unknown=Unknown error while performing your request, report this issue with the following information : \n```\n{}\n```
 chardata_invalid=Data provided are not valid for creating a character
 charlink_conflict=Unable to link {}, the character is already linked to someone. Use `unlink` command or specify `override` parameter with `True` value
+charselect_placeholder=Select your character
